+++ apps/api/filters/feeds.py
@@ -7,6 +7,7 @@ class FeedFilter(django_filters.FilterSet):
     creator_id = django_filters.UUIDFilter()
     catalog_id = django_filters.UUIDFilter()
     title = django_filters.CharFilter(lookup_expr='unaccent__icontains')
+    kind = django_filters.ChoiceFilter(choices=Feed.FeedKind.choices)
 
     class Meta:
         model = Feed