+++ apps/api/forms/catalogs.py
@@ -1,7 +1,8 @@
 from django import forms
-from django_api_forms import Form
+from django_api_forms import Form, BooleanField
 
 
 class CatalogForm(Form):
     url_name = forms.SlugField()
     title = forms.CharField(max_length=100)
+    is_public = BooleanField(required=False)