+++ apps/api/forms/feeds.py
@@ -8,17 +8,18 @@ from apps.core.models import Catalog, Feed, Entry
 
 class FeedForm(Form):
     catalog_id = forms.ModelChoiceField(queryset=Catalog.objects.all())
-    parent_id = forms.ModelChoiceField(queryset=Feed.objects.all(), required=False)
     title = forms.CharField(max_length=100)
     url_name = forms.SlugField(max_length=50)
     kind = forms.ChoiceField(choices=Feed.FeedKind.choices)
     content = forms.CharField()
     per_page = forms.IntegerField(min_value=1, required=False)
     entries = forms.ModelMultipleChoiceField(queryset=Entry.objects.all(), required=False)
+    parents = forms.ModelMultipleChoiceField(
+        queryset=Feed.objects.filter(kind=Feed.FeedKind.NAVIGATION),
+        required=False
+    )
 
     def clean(self):
-        if self.cleaned_data.get('parent_id') and self.cleaned_data['parent_id'].kind == Feed.FeedKind.ACQUISITION:
-            raise ValidationError(_("Cannot have acquisition feed as a parent"))
         if self.cleaned_data.get('entries') and self.cleaned_data['kind'] == Feed.FeedKind.NAVIGATION:
             raise ValidationError(_("Navigation feed cannot have entries"))
         return self.cleaned_data